Create file_search.rs
Let's generate a windows file recursive search program in Rust lang. The program should execute in the command line environment. The program should prompt the user for the search string. Suggest a file name upon completion.

 Below is an example of a simple recursive file search program written in Rust. This program will search for files that contain a specified string in their names, starting from the current directory. The program will prompt the user for the search string and suggest a file name upon completion.

To run this program, you need to have Rust installed on your system. You can install it from the official Rust website.

Now, create a new Rust project with cargo:

cargo new file_search
cd file_search

Then, add this code to the src/main.rs file:

ou can then run the program by executing cargo run in the terminal within the project directory.

This program will print the file paths of the files that contain the search string in their names. After the search is complete, it will suggest a file name for saving the results. Note that the program will only print the file paths to the console and not actually save them to a file. You can extend the program if needed to save the results to a file.