Create file_searchv2.rs
In this modified version, the program now includes a prompt_for_directories function that prompts the user to enter one or more directory paths. The user can enter 'done' when they have finished entering directories. The program then scans the specified directories for files containing the search string in their names.

To run this modified version, replace the code in the src/main.rs file with the above code and execute cargo run in the terminal within the project directory.