Create neosearch.rs
Let's generate a file search program using Rust lang for windows drives. The program should prompt the user for the file or files to be searched and search directories recursively. It should be capable of handling special characters and wildcard strings such as *.pdf which should output all .pdf files, etc. The program should output the search results. Suggest a file name for this file upon completion.

Below is a simple Rust program that accomplishes what you are looking for. It searches for files on Windows drives recursively and supports wildcard characters.

To run this program, you need to have Rust installed. Additionally, you need to use the glob crate for pattern matching, so you'll need to create a Cargo.toml file in the same directory as your file_searcher.rs with the following content:

[package]
name = "file_searcher"
version = "0.1.0"
edition = "2018"

[dependencies]
glob = "0.3.0"

After creating these two files, you can build and run the program by executing cargo run from the command line in the same directory.

Note that searching through entire drives can take a long time, especially if the drives have a lot of files. Also, be cautious when searching and handling files, as it can be easy to accidentally modify or delete important data.