Create sha2search.rs
To generate a hash for each file, you can use the sha2 crate. You'll need to add it to your Cargo.toml file as a dependency:

[package]
name = "neosearch"
version = "0.1.0"
edition = "2018"

[dependencies]
glob = "0.3.0"
sha2 = "0.9.5"

In this refactored version, after the user specifies the output filename, the program will prompt them with "Do you want to generate a hash for each file? (yes/no):". If the user enters "yes", the program will generate a SHA-256 hash for each file it finds and write it alongside the file's path in the output file.