Create multithreaded_hash_and_search.rs
Let's refactor 'multithreaded_neosearch.rs' to generate a hash of each file on a selected hard drive and write that output to a file.  Hash generation should be implemented using parallelism via the Rayon crate. When complete suggest a file name for the completed program.

Sure, let's add the capability to generate a hash of each file. We can use the sha2 crate for generating the SHA-256 hash. We will need to add this to the Cargo.toml file as well. This updated version of the program will perform a recursive search on selected hard drives, generate a hash of each file, and write the file path along with its hash to an output file.

Here is the updated code:

Cargo.toml:

[dependencies]
rayon = "1.7"
glob = "0.3.0"
sha2 = "0.9.8"

I have added a function called generate_file_hash that takes a file path and generates a SHA-256 hash for the file contents. This function is called within search_in_directory, which writes the file path along with its hash to the output file.