Create indexer.rs
Let's generate a Rust program for windows operating systems that uses the Rayon crate parallelism to implement a hashmap index of files on a selected hard disk. The program should ask the user which hard disk or disks to index, what kind of files to index (and should be capable of handling special characters and wildcard strings such as *.pdf , etc), and ask the user how to name the output index. The program should stream the indexing progress upon execution. Suggest a file name for this file upon completion.

Your task can be solved using Rust with the help of a few additional crates:

rayon for data parallelism.
glob for file system globbing which supports patterns like *.pdf.
walkdir for recursively traversing directories.
Here is a sample program:

This program will ask the user which drives to index, the type of files to index, and the name of the output file. It will index all files of the given type in the specified drives and write the index to the output file.

The indexing is performed in parallel for each drive, which could greatly speed up the process if the user has multiple drives.