Create BFDE_v2.rs
Thanks! That's working great. Now I'd like to add some additional functionality: Prompt the user to enter the name of the file to write the crawled websites to. Then prompt the user for the website to crawl.

You can use the std::fs::File and std::io::Write traits to write the crawled URLs to a file. Here's how you could modify your program to include this functionality:

In this updated version of your program, write_urls_to_file is a new function that writes each URL to the file. File::create(filename)?; opens the file (and creates it if it doesn't exist), and writeln!(file, "{}", url)?; writes a URL to the file, followed by a newline.

It's a good practice to handle errors when writing to a file. In this case, write_urls_to_file returns a std::io::Result<()>. If writing to the file fails for any reason, the error is returned and can be printed to the console.

I've also added a prompt at the start of main to get the filename from the user. The filename is then passed to write_urls_to_file after all the URLs have been crawled.