Create hydra.rs
To stream the crawled links, we'll print them out as soon as we crawl them. To write to the file at the same time, we'll open the file in append mode and write the link as soon as we crawl it. Finally, to extract new URLs from the crawled page, we'll parse the HTML and look for a tags, extracting the href attribute as a potential new URL.

Please note, URL extraction in this example is quite naive and for a more robust crawler, URL normalization and validation would be needed. Also, the current code doesn't check if the newly found URL belongs to the same domain.