Rename main.rs to mainv1.rs
In order to exclude duplicate URLs, you need to store the visited URLs in a data structure that allows for quick lookups. One way to do this is by using a HashSet. Below is a refactored version of your code that uses a HashSet to exclude duplicates:

In this refactoring, I added a HashSet called visited_urls which will keep track of all the URLs we've visited so far. Before adding a URL to url_queue, we now check if it exists in visited_urls to avoid duplications. If it's not, we add it to both url_queue and visited_urls. This way, we ensure that each URL is crawled and written to the file only once.