Create main.rs
To refactor the Rust code for Windows systems and implement lazy evaluation of excluded domains, we will modify the urls_to_crawl HashMap to only add URLs to crawl if they do not contain the "facebook" domain. We'll also update the new_urls filtering logic to exclude URLs with the "facebook" domain. We'll add a check for the "facebook" domain in the main loop to avoid crawling URLs with that domain. Additionally, we'll remove the unnecessary Windows-specific user agent string.

Here's the refactored code:

With these changes, the Rust code should now lazily evaluate and skip any URLs with the "facebook" domain while crawling. The crawled URLs will be written to the output file specified by the user. The Cargo.toml file provides the necessary dependencies to build and run the code.